requirements.txt
pycups==2.0.4
ezprinting==0.3.0
python-barcode==0.15.1
upc-item-db-api==0.1.1
Pillow==11.1.0